Create GUI.py
Implementing a visual alternative to your program. The GUI was constructed using tkinter interface. Run using an algorithm found online as so I will not modify your code as of yet. I suggest you study the code and try to implement it in your own python programs. Please let me know if any bugs are found.